code refactoring part 2
base code refactoring was finished
